refactor: extract challenge options
